feat: refactor matching components and utilities for improved structure and readability
